projects/NLP/Chatbots/Case-Study_Restaurant_Chatbot/requirements.txt
absl-py==0.3.0
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
appnope==0.1.0
appscript==1.0.1
APScheduler==3.5.1
asn1crypto==0.22.0
astor==0.7.1
astroid==1.5.3
astropy==2.0.2
attrs==18.1.0
Automat==0.7.0
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
bokeh==0.12.10
boto==2.48.0
boto3==1.5.20
botocore==1.8.50
Bottleneck==1.2.1
bz2file==0.98
certifi==2018.4.16
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
coloredlogs==10.0
colorhash==1.0.2
conda==4.4.7
conda-build==3.0.27
conda-verify==2.0.0
ConfigArgParse==0.13.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.0.3
cycler==0.10.0
cymem==1.31.2
Cython==0.26.1
cytoolz==0.9.0.1
dask==0.15.3
datashape==0.5.4
decorator==4.2.1
dill==0.2.8.2
distributed==1.19.1
docopt==0.6.2
docutils==0.14
duckling==1.8.0
en-core-web-md==2.0.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fakeredis==0.10.3
fastcache==1.0.2
fbmessenger==5.2.0
filelock==2.0.12
Flask==1.0.2
Flask-Cors==3.0.4
future==0.16.0
gast==0.2.0
gensim==3.4.0
gevent==1.2.2
glob2==0.5
gmpy2==2.0.8
graphviz==0.8.3
greenlet==0.4.12
grpcio==1.14.0
h5py==2.7.1
heapdict==1.0.0
html5lib==0.9999999
humanfriendly==4.16.1
hyperlink==17.3.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
imbalanced-learn==0.3.1
incremental==17.5.0
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
JPype1==0.6.2
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
Keras==2.1.6
klein==17.10.0
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lxml==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
mattermostwrapper==2.1
mccabe==0.6.1
mistune==0.8.3
mpmath==0.19
msgpack==0.5.6
msgpack-numpy==0.4.3.1
msgpack-python==0.5.4
multipledispatch==0.4.9
murmurhash==0.28.0
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.4
nose==1.3.7
notebook==5.4.0
numba==0.35.0+6.gaa35fb1
numexpr==2.6.2
numpy==1.14.3
numpydoc==0.7.0
odo==0.5.1
olefile==0.44
openpyxl==2.4.8
packaging==17.1
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.3.1
pathlib==1.0.1
pathlib2==2.3.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.4.0
pickleshare==0.7.4
Pillow==4.2.1
pkginfo==1.4.1
plac==0.9.6
ply==3.10
preshed==1.0.1
prompt-toolkit==1.0.15
protobuf==3.6.0
psutil==5.4.0
ptyprocess==0.5.2
py==1.4.34
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
pydot==1.2.4
pydotplus==2.0.2
pyflakes==1.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.6.4
pykwalify==1.6.0
pyodbc==4.0.17
pyOpenSSL==17.2.0
pypandoc==1.4
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.1
python-crfsuite==0.9.5
python-dateutil==2.6.1
python-telegram-bot==10.1.0
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
redis==2.10.6
regex==2017.4.5
requests==2.18.4
rope==0.10.5
ruamel-yaml==0.11.14
ruamel.yaml==0.15.37
s3transfer==0.1.13
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.8
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.11.0
sklearn-crfsuite==0.3.6
slackclient==1.2.1
smart-open==1.5.7
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
spacy==2.0.11
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.13
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tabulate==0.8.2
tblib==1.3.2
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.3
toolz==0.8.2
tornado==5.0
tqdm==4.23.3
traitlets==4.3.2
twilio==6.14.0
Twisted==18.7.0
typing==3.6.4
tzlocal==1.5.1
ujson==1.35
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.48.0
Werkzeug==0.14.1
widgetsnbextension==3.0.2
wrapt==1.10.11
xgboost==0.71
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.4
xlwt==1.2.0
zict==0.1.3
zomatopy==1.0.10
zope.interface==4.5.0
